TensorBoard profiling example (ex5)
Users has requested information and examples of debugging
TensorFlow applications. TensorBoard is a visualization toolkit
bundled with TensorFlow. TensorBoard allows basic profiling and
tracing using the TensorFlow Profiler. This example aims to illustrate
how the module system, TensorFlow with TensorBoard callback, and TensorBoard
UI can be used in conjunction to profile and trace a simple ML-model.