Update README.md
Removed line about setting "use GPU" since this was fixed in the script - the arrays also need to be GPU Arrays, so it would be a bigger and less intuitive change than is suitable. It's probably better that it tries to use the GPU by default and complains if it can't.